Bugged UserListView
having trouble getting a list of users working. the list needs organizers and entrants to be part of it, since both can act as entrants, so the list needs to be a list of users to cover both. However, adding an entrant to the user list makes the program crash. I'm done the UI though so I'm gonna call that someone else's problem for now